+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Add `ContainerCreateOptsBuilder::image_arch`
 * Add `ContainerCreateOptsBuilder::manage_password`
 * Add `ContainerCreateOptsBuilder::mounts`
+* Add `ContainerCreateOptsBuilder::passwd_entry`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync